CA-123658: Fix up legacy QEMU suspend image restore
Older versions of QEMU seem to generate larger suspend images and in some
versions of the product were greater than the 8k which had been allocated for
the buffer. Added a fix to realloc this buffer once we know how big the image
is (which is contained in some metadata in the legacy QEMU save signature).

Also incorporated style improvements based on comments from Andrew Cooper.