CA-106594: enable lcoal irqs when testing the NMI watchdog
wait_for_nmis() runs for 10 ticks with local interrupts disabled,
which is quite a long time.

We occasionally see serial output (particularly with virtual serial
ports such as HP's iLO) stopping on the "Testing for NMIs" line.  We
think the long delay in handling the serial interrupt is the cause.

Also, reduce the verbosity of the "Testing NMI Watchdog" line if it is
all successful.