CA-95221 - Make IOMMU Dom0 PCI passthrough default to on
The current Xen default is to create a 1:1 mapping of all physical RAM for
the IOMMU with Dom0 domain context using IOMMU pages.

This results in a IOMMU lookup for all DMA activity which could lead to
IOTLB thrashing in a heavy IO in Dom0 scenario.

The IOMMU passthrough option bypasses looking up the IOMMU for the DOM0
context and avoids the IOMMU lookups and any potential IOTLB thrashing.