Remove "xen-npt-detect-barcelona.patch"
The origins of this patch have sadly been lost.  It was introduced in 2007,
but lacked a ticket number, and the cpuid test in it does not match any
documented processors from AMD.

It is possible it applied to a XenRT machine at the time, but XenRT has no
working hardware today which matches the test.

It is also possible that it refered to AMD Erratum #383, for which Xen has now
gained a proper workaround for.

Either way, Barcelonas are ancient and test boxes/SDVs are unlikely to be
found in production environments.  We can always reinstate the patch if people
complain, suggest "no-hap" on the command line as a workaround, or suggest
buying some real production hardware.