CP-10412: xen: increase default number of PIRQs for hardware domains
The default limit for the number of PIRQs for hardware domains (dom0)
is not sufficient for some (x86) systems (e.g., the new NetScaler
SDX).

Since the pirq structures are individually and dynamically allocated,
the limit for hardware domains may be increased to the number of
possible IRQs.

The extra_guest_irqs command line option now only allows changes to
the domU value.  Any argument for dom0 is ignored.