CA-157766: Handle shadow ops returning -EBUSY
A shadow op may return -EBUSY if there is another pending preempted
hypercall (issued by another task).  The op should be retried in this
case.

This may fix a problem where migration failed because log dirty could
not be enabled, but in case it doesn't log all the return values from
the 3 different shadow ops.