CA-162192: Fix getting EFI variable list on some systems
Copy the entire output buffer to the guest because it may contain data beyond
the output size that the firmware requires on a subsequent
GetNextVariableName() call (e.g. a NULL character).

The spec requires that on each call, "the previous results" be passed in.

Without this change, the following sequence would occur:
GetNextVariableName: in \0, out AdminPw\0, size 7
GetNextVariableName: in AdminPw\0, out UserPw\0, size 6
GetNextVariableName: in UserPww\0, NOT FOUND