CA-164198: efi: Reserve SMBIOS table region when EFI booting
Some EFI firmware implementations may place the SMBIOS table in RAM
marked as BootServicesData, which Xen does not consider as reserved.
When dom0 tries to access the SMBIOS, the region is not contained in the
initial P2M and it crashes with a page fault. To fix this, reserve the
SMBIOS region.

Also, fix the memcmp check for existence of the SMBIOS and the DMI
checksum calculation.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>