CA-159458: credit: don't overflow dom0's weight
If there are several runnable/running VCPUs with the maximum weight,
the calculation of the weight for the priviledged domain (dom0) may
overflow a uint16_t.  This results in a bogus weight and triggers a
BUG_ON() later on.

Clamp dom0's weight to the maximum.  Also use unsigned ints in
__csched_adj_adjust_privdom_weight().