CP-11509: efi: Reallocate memory map if ExitBootServices() fails
If calling ExitBootServices() fails, the memory map size may have
increased, so determine the new size and reallocate the memory map
before calling GetMemoryMap() again.

This was discovered when trying to do an iSCSI boot on xrtuk-11-09. For
whatever reason, if the UEFI firmware sets an iBFT, ExitBootServices()
always fails on the first call.