CA-177409: increase xenstore maxwatch quota
It is quite reasonable to allow VMs to have more than 100 watches. For example,
since a VM requires a watch on each device's backend state field, a VM with many
VBDs or VIFs could exhaust this number.

The previous limit of 100 was arbitrary. The memory overhead of permitting a
larger number of watches is minimal.