CA-181640: xenguest: add RMRR regions to memory map (if needed)
If a HVM guest as a PCI device assigned to it and the PCI device has
RMRRs, expand the MMIO hole to cover these regions and add the RMRRs
to the e820 memory map.  This allows PCI devices with RMRRs to be
assigned to HVM guests again (fixing a regression introduced with Xen
4.6).

Since the MMIO hole is expanded, 32-bit guests may see less memory
than in Cream.  Fixing these requires changes to domain build to
populate the physmap using a more complex memory map than just two
low/high RAM regions.

As the toolstack does not yet provide the correct -pci_passthrough
option, as a temporary hack check for Intel graphics passthrough via
xenstore and use a hardcoded 0:2.0 device in this case.  Once the
toolstack uses the -pci_passthrough option this hack can be removed.