CA-186953 - Add percpu rwlocks and implement for grant table
Add a generic percpu rwlock implementation. Percpu rwlock's reduce overhead
for taking read locks by using a per cpu variable for the read lock itself.

Traditional rwlocks use compare swap instructions on a specific data location
to track the total number of readers. This approach does not scale on
multisocket systems due to the high overhead of compare swap instructions. The
compare swap also exhibits higher overhead when the read data location is
heavily used and often the comparison fails, resulting in a retry of the
expensive compare swap instruction. Percpu rwlocks store the indication of a
reader using the lock in a per cpu area. A writer taking the lock has to signal
it's intention via a global variable so that readers use the standard read lock
function block on it and the writer then has to poll the per cpu area's

Xen does not require recursion of the rwlock itself and so the percpu rwlock
does not implement this.

Grant table users of the rwlock have been converted to the percpu
rwlock. Converting these locks allows for multiqueue performance scalability
between VM's on multi socket hosts. On a 2 socket systems performance goes
from 10 gbit/s to 40 gbit/s.