CA-189353: Fix read_cr0 intercept when smap fixup is enabled
If a 32 bit PV guest is running under smap fixup mode, CR0.AM is always
disabled when running in the context of the guest. Ensure that the bit
is not set when when intercepting a read of CR0.

This bug can be seen when a guest does a read of CR0 followed by a write
of CR0 which Xen fails with "Attempt to change unmodifiable CR0 flags.",
because the AM bit was unexpectedly set in the result of the read of
CR0.