CP-14765 - Use atomic count of VCPU's on PCPU runqueue to avoid taking spinlock
Credit 1 smp load balancing tries to find work on other PCPUs by looking through
their runqueues. It takes the PCPU runqueue spinlock in order to safely process
the list. This causes lock contention on the PCPU runqueue and cache line boucing.

Use an atomic count of items on PCPU runqueue so that VCPU are not stolen from a
PCPU in an under committed situation and so that the runqueue spinlock is not
taken needlessly.

This should help performance as it reduces cache line bouncing of the spinlock
itself and it should help improve keep the caches hot. Additionally, it may help
improve turbo behaviour on Intel processors as cores will be able to idle more
often.