CA-198942: cpuid: handle inverted CPUID bits
Although featuresets have inverted bits to indicate features that are
removed (e.g., FPCSDS), the CPUID leaf values must be use the hardware
sense (or the guest sees the wrong sense).

When loading and validating a CPUID policy, make sure to properly
flip any inverted bits.

The end result is the FPCSDS bit is the same as in bare-metal, which
is actually what we want (since the hardware behaviour cannot change).
This doesn't highlight that the "inverted" featureset concept is not
what we want.  Instead we should have an "always matches hardware"
featureset, which toolstack can exclude when levelling.