CA-196985: x86/fpu: workaround inability to restore full FPU exception state
There is no combination of FPU restore instructions that can restore
the full 64-bit FIP/FDP and FCS/FDS registers in the FPU.

On hardware that does not have the FPCSDS feature (where the hardware
does not save the FCS/FDS registers but writes 0x0000 instead), this
means that the guest can see inconsistent FPU state.

This normally isn't a problem because FCS/FDS is not interesting in
64-bit mode.  However, Microsoft's Driver Verifier adds a check around
the driver's interrupt handler that checks the FPU (XSAVE) state does
not change across the handler.  64-bit Windows 7 will use a 32-bit
XSAVE instruction for this, saving FCS/FDS.  Xen will do a 64-bit
XSAVE/XRSTOR which clears FCS/FDS.  Windows notices this difference
and raises a 0x3D BugCheck.

We workaround this by allowing the toolstack to request that Xen
always does a 32-bit XSAVE for a particular domain.  The HVM parameter
HVM_PARAM_X87_FIP_WIDTH can be set to 8 (64-bit XSAVE), 4 (32-bit
XSAVE) or 0 (existing automatic detection).

The "platform:x87-fip-width" VM parameter has been added. Xenguest
will read this parameter and set the HVM parameter if the guest is
still in automatic mode.  If the current FIP width is non-zero,
xenguest does not override it because on hardware with FPCSDS, the
default value of 8 is always the safest option.

XenRT will be expected to set this VM parameter when running the
Driver Verifier tests.

We could have used the write to the viridian Guest OS ID MSR to
identify a Windows guest and set the FIP width to 4, but this could
potentially lose important exception state while the guest is handling
a floating point exception.