CA-201372: x86: don't flush the whole cache when changing cachability
Introduce the FLUSH_CACHE_BY_VA flag to flush_area_mask() and friends
to say that it is safe to use CLFLUSH (i.e., the virtual address is
still valid).

Use this when changing the cachability of the Xen direct mappings (in
response to the guest changing the cachability of its mappings). This
significantly improves performance by avoiding an expensive WBINVD.

e.g., A set_memory_wc() call in Linux:

before: 4097 us
after:    47 us