CA-204478: Set mmio_relax for dom0 by default
Set mmio_relax on for dom0 to fix an issue where the dom0 kernel fails
due to an unexpected UC mapping.

Specifically, when using UEFI, mcelog maps 0xec000 intending for the
page to be WB. Since Xen considers this address to be MMIO, due to
XSA-154, Xen keeps the page type as UC. When mcelog daemonizes and
forks, the kernel tries to copy the VMA. It notices that the actual page
type is UC when it is meant to be WB and fails the fork. (It is
hardcoded in Linux that the low ISA memory region (which contains
0xec000) is always WB.)