CA-211532: Intercept and handle reads of MSR_VMX_VMFUNC
The case statement in vmx_msr_read_intercept() which dispatches on the
MSR being read did not include MSR_VMX_VMFUNC, so reads of this MSR were
passed through by default, regardless of whether VMX was enabled for
the guest.   The correct long term solution for this is to whitelist
supported VMX MSRs rather than blacklisting them, but for now we will
just extend the case statement and hope that when new VMX MSRs are added
someone will remember to extend it again.

Signed-off-by: Euan Harris <euan.harris@citrix.com>