livepatch: Ignore duplicate new symbols
For XenServer, we build cumulative live patches and always use the
replace operation so that there is only ever a single active live patch.
It is simpler to build, distribute, and manage patches like this.
However, this may cause duplicate new symbol errors. Because of the
lookup order (hypervisor first) and the fact that live patches will only
ever be built as diffs against the base hypervisor (and not another live
patch), it should be safe to ignore these errors.