CA-220429: xen/console: Set the default log level to INFO for release builds
With the default log level as it currently is, the messages for applying
and reverting live patches do not get logged as was originally intended.
By setting the default log level to include INFO, these messages are
printed to the console.

While this change does increase the boot verbosity a little, it doesn't
appear to cause additional log messages to be displayed during normal
operation.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>