CA-219663: Save CR2 in R9 if copy_to/from_guest fails
If copy_to_guest() or copy_from_guest() fails (because the guest
address is not accessible), then save the CR2 (linear fault address)
register value into the guest's R9 register so the guest can tell
which address caused the fault.

This requires a change to the guest kernel to mark R9 as a output when
performing a hypercall.  Thus, we only do this if we're in a hypercall
and only for dom0.

This change should allow us to get some useful information about some
unexpected -EFAULT errors on some hypercalls from xenguest.