CP-19488: Fix spec files and add configuration enabling CONFIG_GCOV
Fix the spec file following the gcov updates and add an example
hypervisor build configuration file that enabled CONFIG_GCOV.

We don't produce a build with coverage enabled since its not yet clear
how this would be consumed and what else it would require in it.
e.g., Do the *.gcno files need to be packaged?  How will the matching
source be obtained?