Move to the latest stable-4.7 (with XPTI)
Upstream version of XPTI (Meltdown mitigation) is incompatible with
out-of-tree version of Spectre mitigation that currently included in XS
as a rather large technical debt. This change removes this debt and
uses upstream version of Spectre mitigations from stable-4.7 branch.

Base version is still RELEASE-4.7.4 because in order to change it,
one has to change the spec file in xenserver-specs (or overrides) which
is inconvenient. So instead, all the patches from stable-4.7 are added
as backports.

Signed-off-by: Sergey Dyasli <sergey.dyasli@citrix.com>