CP-24243: Remove IGD passthru hack
Remove the IGD passthru hack and pass through the IGD as a normal
device. The graphics driver in Windows is now able to use it in UPT
mode without any legacy baggage.

Move the ROMBIOS patch which was indicated as required for GVT-d into
the GVT-g specific section and add a comment describing why it is
needed.