CA-286032: Move mxGPU workaround from hvmloader to xenguest
hvmloader contains a workaround for mxGPU cards to set the MMIO hole
size to 2G. But this is incompatible with QEMU upstream since xenguest
communicates the size of the MMIO hole to QEMU. Consequently, mxGPU
doesn't work with QEMU upstream because it uses a different MMIO hole
size. To fix this, move the workaround into xenguest instead which
allows QEMU upstream to use the correct size.