CP-27581: Make MMIO hole size key persistent across migrations
Toolstack should preserve this key on migrations which makes possible
to migrate with variable MMIO hole size (as the key is later used for
QEMU initialization on the destination host). This protects a VM from
unpredictable modifications in device model configuration like
introduction of a new migratable virtual device or third-party changes
in device emulators (i.e. DEMU) - we don't need to guess it like
before.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>