CA-290052: Don't allocate M2B for Dom0 translations in PV-IOMMU
As soon as vGPU makes a request to pin a page for DMA using
IOMMUOP_lookup_foreign_page, PV-IOMMU allocates M2B entry
for the requested pages. The initial idea was to track page
references that way as we get the page every time we are
making a pin request to avoid security problems with ballooning.
But that seems unnecesary as we trust vGPU host driver with
making an unpin request (IOMMUOP_unmap_foreign_page) for every
pin request that it makes.

As M2B tracking occupies significant amount of Xen heap space
(~32MB per 1GB) and breaks XAPI memory accounting let's skip
M2B allocations for Dom0 requests. However, that has an
unfortunate drawback of leaving dangling page references in
case of a vGPU driver fault.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>