CA-290664: Remove M2B and reinstate flag based memory tracking in PV-IOMMU
Relying on the host drivers is not enough for tracking memory as
they often don't track it themselves. Replacing M2B with page flag
based scheme is also zero overhead and additionally covers the cases
where a host driver is not reliable (old NVIDIA drivers and GVT-g).

The downside is that this scheme is only available for Dom0 as only
one mapping for a page can be tracked at a time. This shouldn't be
a problem for our internal use-cases.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>