CA-319984: Increase default max grant frames for VMs
VMs with high numbers of vCPUs and VIFs hit the max grant frames limit.
Furthermore, with Xen 4.11 this limit for VMs is no longer configurable
on the Xen command-line as it was previously.

Increase the default to 64 frames for VMs. This is sufficient for the
maximum XenServer configuration limits:

7 VIFs * (256 rx entries + 256 tx entries) * 8 queues max * 8 bytes / 4096
= 56 pages which is < 64.

Allow this value to be overridden by setting a platform flag to avoid
having to recompile xenguest should the administrator want to go beyond
supported limits.