Fixes to local quirk handling
The quirk-*.patch patches are quite problematic.

The 3 backports are upstream cleanup for DMI handling, in preparation to
simplify quirk-hp-gen8-rmrr.patch substantially.  On Gen8 systems, don't
bother reading the PCI_CLASS_DEVICE unless we're going to look at it.

For quirk-pci-phantom-function-devices.patch, the logic is missing breaks in
the outer switch statement, meaning that for a matching vendor, we fall
through and read PCI_DEVICE_ID up to 3 times.

Rearrange the logic to only read PCI_DEVICE_ID once, and break such that we
don't consider mismatched vendor/devices pairs for having phantom functions.
Fold single-entry switch stamtents into an if().

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>