CA-337060: Keep reserved regions identity mapped in Dom0
Some machines are unhappy when E820 regions like reserved and ACPI are
not identity mapped. For PV-IOMMU it's only necessary to have regular
RAM regions mapped at offset so keep all the others mapped as they
were before according to boot options like "map-inclusive" and
"map-reserved" but only if MFN address is below 1-1 map offset.

This does correspond to the end result of previous PV-IOMMU mappings.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>