CP-32980: Xen security policy for Hitachi storage driver domains
A domain other than Dom0 cannot act as a driver domain due to silo mode restrictions.
We want to lift this restriction for storage driver domains, as used by Hitachi SDS (REQ-802).
We won't support driver domains in general though, it still activated by an otherwise experimental other-config field in xapi.

In the long run a Flask policy would be preferable, but Flask is not ready yet.
Meanwhile hardcode a magic SSID that gives a domain just enough extra privileges to ignore silo mode restrictions.
This is not the same as is_privileged, because that would grant far more privileges. BitDefender needs that,
but a Hitachi driver domain doesn't.

The toolstack will pass the magic SSID when a VM is a driver domain.

Patch was written by Andrew Cooper, so I assume:
Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>