CA-337004: Allocate a reverved region for vGPU VRAM within MMIO hole
The current location at 0xFF000000 vGPU is using to allocate memory for
VRAM is not appropriate since it overlaps with OVMF startup position just
below 4G. That eventually results in incorrect mapping being created for
VRAM in case NVIDIA SR-IOV device is used.

A new region is chosen so as to not overlap with existing RMRR regions
at or below 0xFB000000. It's marked as reserved in E820 which makes
hvmloader to think of it as an RMRR and map PCI BARs accordingly
to avoid overlaps. The final address is then saved to xenstore
the same way MMIO hole size is currently saved for a VM - that allows
DEMU to get the location where it's allowed to allocate memory on
boot and where to find it after suspend/resume.

VRAM reserved region size is currently hardcoded to 16M since NVIDIA
drivers are not able to work with anything else but an additional key
is saved to xenstore for the future.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>