CA-323738: Pick only online vCPUs as i8259 default target for timers
When kdump is invoked on a vCPU != 0 the next kernel will try to test
PIT timer interrupts that on real hardware are supposed to be routed
through IOAPIC with PIC disabled. With vPIT series applied it's now
possible to correctly route these interrupts through IOAPIC. But vPT
susbystem still assigns vPIT timer to a global timer vCPU target which
now equals to i8259 target. That means if this vCPU target is suddenly
offline (what is actually happening on kdump) then timers are not
processed at all.

For now instead of always choosing vCPU0 as a default i8259 target -
pick the first online vCPU instead. That will guarantee timers are
kept processed even when vCPU0 is going offline on kdump. Additionally,
recalculate i8259 target on vCPU offline.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>