CA-343455: Only map incorrectly typed EFI reserved regions below 4G for RS
The existing workaround seems to cause upcoming HPE systems based on IceLake
CPU to trip over a gigantic reserved region at the top of EFI memory map
throwing a BUG_ON on page table allocation. Having one is not strictly
forbidden and most operating systems would cope well with it so our
behavior is inadequate here.

I tend to keep the existing workaround in place as there could be still
systems that need it but would like to restrict it to only map those regions
below 4G which is in line with a similar workaround enabled by default in
IOMMU code for mapping reseved regions into Dom0 context. It's unlikely
to have any RS memory above 4G anywhere.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>