CA-340135: Give headroom for virtual HW BARs in MMIO hole
If the expected virtual hardware MMIO resource requirement is not taken
into account initially, there might be a situation where the final size of
MMIO hole is not enough to accomodate a passed through PCI device.
In case of 64 bit guest that usually goes unnoticed as a BAR of the PCI
device is simply reallocated above 4G. For 32 bit guest in contrast that
poses an issue as Windows for instance doesn't have PAE support enabled
and therefore doesn't have access to the area above 4G.

Take a chance and tidy up the PCI resource gathering loop a little.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>