CA-350524: Reduce number of entries in MADT to 64 to appease new Windows kernels
Upon inter-op pre-release testing of Windows Server vNext (build 20270)
it was revealed that SMP now fails to initialize with Viridian enabled.
That turned out to be due to the fact Windows became more strict in applying
CPU limits inherent to our Viridian interface and the way it now calculates
number of CPUs available:

It appears that the limit of 64 that we currently advertise according to TLFS
is now applied to the number of MADT entries regardless of their status
(enabled or disabled).

We could remove the limit of 64 VPs by fixing Viridian implementation of Xen
by backportig CPUID leaf changes and ExProcessorMasks extensions. Or we can
simply reduce the number of MADT entries to 64 given this is a local quirk
and we do not support more than 32 vCPUs per VM now in the product (room still
left for a bump to 64).

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>