CA-356226: Fix incoming VMs which saw MPX
In Xen, MPX being available needs to be accounted for when calculating the
maximum possible features the VM can use.

In xenguest, construct_cpuid_policy() needs calling ahead of
xc_domain_restore() so the policy (with non-default MPX) can be constructed
appropriately for the VM before the incoming register state is audited.

This in turn involves fixing the HVM_PARAM_PAE_ENABLED mess.  Most bugfixes to
this have been backported, with just one remaining to allow us to treat PAE as
just a plain feature like everything else in the featureset.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>