CP-38517: Create a subpackage for building live patches
Create a sub package that contains the required files for later
building live patches against this base version. It contains the
prepared sources files, config, and built xen-syms file. While the
contents are available in other built RPMs, they are either not
installable as build dependencies or not easily installable in parallel
for building against multiple base versions.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>