Simplify CPUID rescanning massively
There are at least three bugs:
 * update_x86_caps() clobbers boot_cpu_data which is cpu_data[0] before
   merging.  As a consequence, the logic returns noop even when something has
   changed.
 * If a timeout occurs, 'static bool error' is never reset, meaning that no
   subsequent update will complete, even if the system is quiet enough now.
 * Because the update doesn't account for forced_caps[], RTM_ALWAYS_ABORT in
   particular gets lost if an update is performed.

This is all to do with CPUID, and not really to do with speculation, so
shouldn't live in spec_ctrl.c.  However, to avoid losing features, the update
logic must take account of forced_caps[], meaning that the logic needs to be
in cpu/common.c

There is absolutely no need for the use of stop_machine_run(), and by
extention, continue_hypercall_on_cpu().  Rewrite it from scratch, using
regular IPIs.

The only other aspect that matters is to exclude new domain creation while
we're updating the system CPU policies, so take the domctl lock too.

This simplification also removes the (mis)use of EFAULT for "misc error".

Example of it working:
  (XEN) New CPUID features detected
  (XEN)   featureset word 9, new 00000800

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>