CP-40031: vTPM: Access the PPI region using IO ports
Varstored exposes the PPI region as an index/data IO port pair.
Update the ACPI TPM code to use this instead of the memory region
exposed by QEMU.

While ASL does support access via index/data registers in an elegant
manner (using IndexField), it requires that the index is aligned to the
size of the data field. In this case, some 32-bit accesses are not
aligned so IndexField can't be used. Instead, a few wrapper functions
are defined to make access easier.

At the same time, update the code to use the operation flag values as
defined in edk2.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>