CP-40617: Build the hypervisors in the %build phase
Xen's build system can't currently cope with the RPM CFLAGS/LDFLAGS, so build
the various hypervisors after the tools, and unsetting the variables.

We can't put artefacts directly into the buildroot, so stash them in a subdir
in the source tree.

Build the Shim Xen's in the same manner as the main Xen's.  This should be
equivelent to how they were built before.

In all 3 cases (build_xen, install_{xen,shim}), provide a function to avoid
opencoding the runes.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>