CP-45175: Allow overriding the OVMF firmware at runtime
Read from the per-VM platform/ovmf-override key to use a different build
of OVMF from the compile-time default. This is useful to boot a VM with
the debug build of OVMF without disturbing the system - currently this
requires changing symlinks or renaming files.

For security reasons, ovmf-override may only specify a regular file in
the compiled-in OVMF firmware directory (typically /usr/share/edk2 on
XenServer). The built-in path must also point to a regular file.

At the same time, change it so that xenguest fails when it is told to
load OVMF but the OVMF path was not set during configure.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>