CA-375021: Declare setup and dependencies for building live patches
Ship a new script with the lp-devel package, "prepare-build", that is
used to set up the environment to choose the correct compiler etc. so
that it can be used during future live patch builds. This is necessary
since different versions of Xen may be built with different toolchains.

Also add the build dependencies of the main hypervisor build as regular
dependencies of the lp-devel package so that it can be used to pull in
the correct dependencies during the live patch build since different
builds of Xen may require different dependencies.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>