CP-46857: Remove hvmloader vCPU limit
This is a revert of the XSI fix introduced in CA-350524. From the recorded
history it appears the problem was present and fixed in a pre-release version
of Windows. The presence of this patch prevents launching any VM with more than
64 vCPUs, but non-Windows VMs work just fine.  No VM can be started with this
many vCPUs anyway because the templates preclude it, so remove this piece of
tech debt.

Signed-off-by: Alejandro Vallejo <alejandro.vallejo@cloud.com>