CA-399561: Fix name of reboot-required.d stamp file to fix livepatching
f537f07800a78cf2cc25887e4ed6e7f0fa2b0fd9 altered the name of the stamp file in
reboot-required.d from `%{name}/%{version}-%{release}` to
`%{name}/%{version}-%{hv_rel}`. This means it went from e.g.
`4.13.5-9.43.xs8~2_1` to `4.13.5-9.43`. It is believed this was done as part of
a bulk find/replace operation.

The stamp file in reboot-required.d is used to facilitate the livepatching
functionality - if the livepatch successfully applies, the stamp file is
removed, indicating to the toolstack the reboot is not needed.

The logic in livepatching currently expects the full dist tag to be used, and as
such the prior change has broken this functionality, and the toolstack continued
to indicate a reboot is required.

Restore the original filename to fix this issue.

Signed-off-by: Alex Brett <alex.brett@cloud.com>