CP-53418: Scan for microcode by default
Backport a patch to add a Kconfig option to allow setting the default
ucode scan behaviour and set it to scan by default. This ensures that
up-to-date microcode is used at installation time and during upgrades.

Additionally backport a patch (and dependencies) to avoid scan=true and
an index being set at the same time. This is more important now that
scan may be implicitly set to true.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>