CA-414083: account for PCI MMIO used memory in hvmloader
xenguest vGPU can steal the end of the PCI MMIO hole and set it as reserved
in the memory map for vGPU usage.  This conflicts with the code that
attempts to position the Xen PCI platform device BAR as WB, as it also uses
the end of the PCI MMIO hole to do so.

Adjust hvmloader to account for xenguest having possibly moved the
usable PCI MMIO windows end.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>