Implement extendable XEN_DOMCTL_get_runstate_info interface
Having xc_runstate_info_t be the boundary between the dynamic libxenctrl.so
and the statically-linked xenctrl_stubs.cma causes stack corruption if
xc_runstate_info_t grows.  To fix this, we need to rearrange the boundary.

For Xen and libxc, switch fully over to struct xen_domctl_runstate_info_ext,
which can be safely extended owing to the enclosing domctl union.

For the Ocaml side, create new types and stubs which will have extra fields
added to them in the immediately following patches.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>