CP-308091: Mark .rodata as writable in xen.efi
During early boot, Xen might write to __ro_after_init variables that
live in .rodata which will cause a page fault if the loader has set the
page permissions according to section characteristics in the PE file.
Therefore, mark .rodata as writable in the PE file. It will still be
marked as read-only in init_done().

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>