CA-408856: Make sure device are quarantined on deassign
If lockdown mode is enabled do not allow device to go directly
to dom0.
This makes possible to safely reset a device passing to the
quarantine domain and resetting the card.
As Xen does not handle the reset leave this job to dom0 that
will track the device and reset.

Signed-off-by: Frediano Ziglio <frediano.ziglio@citrix.com>